% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vic4to5.R
\name{vic4to5}
\alias{vic4to5}
\title{VIC4to5}
\usage{
vic4to5(soilfile)
}
\arguments{
\item{soilfile}{the VIC-4 soilfile}
}
\value{
NetCDF input file for VICWUR
}
\description{
Convert VIC version 4.x ascii input data to VICWUR (VIC version 5.x) NetCDF input data
}
\examples{
vic4to5()
}
\author{
Wietse Franssen \email{wietse.franssen@wur.nl}
}
